requirements.txt
certifi==2018.11.29
Django==2.1.7
djangorestframework==3.9.1
djangorestframework-simplejwt==4.0.0
pipenv==2018.11.26
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2018.9
virtualenv==16.4.1
virtualenv-clone==0.5.1
